+++ docs/CHANGES.md
@@ -1,8 +1,9 @@
 
-##### 0.3.0  -  todo in progress
+##### 1.0.0
 
- 
+Implemented
 
-
+- Jackson 2 module for Ser/deserialization of Dict and LocalizedString
+- basic utility class for working with Jackson